requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.12
jinja2==3.1.4
reportlab==4.2.5
aiofiles==24.1.0
certifi==2024.8.30
